# packages/contracts/openapi.json
{
...
  "openapi": "3.1.0",
  "info": {
    "title": "Bookflow API",
    "version": "0.0.0",
    "description": "Generated from the Zod route schemas in apps/api. Never hand-written (ADR-014, ADR-025)."
  },
  "components": {
    "schemas": {
...
      "CreateTeamMemberRequestInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Required. Trimmed. 1–200 characters after trimming."
          },
          "role": {
            "type": "string",
            "maxLength": 100,
            "description": "Job title, e.g. \"Senior stylist\". NOT an authorization role."
          },
          "about": {
            "type": "string",
            "maxLength": 2000
          },
          "photoUrl": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "name"
        ],
        "description": "Someone to add to the team."
      },
      "UpdateTeamMemberRequestInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200,
            "description": "Required. Trimmed. 1–200 characters after trimming."
          },
          "role": {
            "type": "string",
            "maxLength": 100,
            "description": "Job title, e.g. \"Senior stylist\". NOT an authorization role."
          },
          "about": {
            "type": "string",
            "maxLength": 2000
          },
          "photoUrl": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri"
          },
          "position": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "description": "The fields to change. At least one."
      },
...
      "TeamMember": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "name": {
            "type": "string"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "about": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "photoUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "position": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "id",
          "name",
          "role",
          "about",
          "photoUrl",
          "position"
        ],
        "additionalProperties": false,
        "description": "Someone a client can book with."
      },
...
    }
  },
  "paths": {
...
    "/v1/me/business/team-members": {
      "get": {
        "operationId": "listMyTeamMembers",
        "summary": "The caller's team",
        "tags": [
          "team"
        ],
        "description": "In display order, ties broken by creation time. An account with no business gets an empty list.",
        "responses": {
          "200": {
            "description": "The salon’s team, in display order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TeamMember"
                  },
                  "description": "The salon’s team, in display order."
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createTeamMember",
        "summary": "Add a team member",
        "tags": [
          "team"
        ],
        "description": "One name field (ADR-005): team members are content records, unlike the owner’s own account. `role` is a job title, never an authorization role. Names are not unique — two stylists may share one.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTeamMemberRequestInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMember"
                }
              }
            }
          }
        }
      }
    },
    "/v1/me/business/team-members/{memberId}": {
      "patch": {
        "operationId": "updateTeamMember",
        "summary": "Change a team member",
        "tags": [
          "team"
        ],
        "description": "Every field is optional and at least one must be present. A member who is not the caller’s is indistinguishable from one who does not exist.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTeamMemberRequestInput"
              }
            }
          }
        },
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "memberId",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TeamMember"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTeamMember",
        "summary": "Remove a team member",
        "tags": [
          "team"
        ],
        "description": "Hard delete (ADR-036). ADR-006 snapshots the team member onto every booking, so removing one never rewrites a booking they took.",
        "parameters": [
          {
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            },
            "in": "path",
            "name": "memberId",
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Default Response"
          }
        }
      }
    },
...
  }
}